# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AhmedHossam777_go-mongo_internal_dto.CreateCourseDto:
    properties:
      course_name:
        maxLength: 100
        minLength: 2
        type: string
      instructor_id:
        type: string
      price:
        type: integer
    required:
    - course_name
    - instructor_id
    - price
    type: object
  github_com_AhmedHossam777_go-mongo_internal_dto.CreateUserDto:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_dto.UpdateUserDto:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_models.Course:
    properties:
      course_name:
        type: string
      id:
        description: If ID is empty, don't include it in JSON/BSON
        type: string
      instructor_id:
        type: string
      price:
        type: integer
    type: object
  github_com_AhmedHossam777_go-mongo_internal_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_AhmedHossam777_go-mongo_internal_models.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A RESTful API for managing courses, users, and authentication built
    with Go and MongoDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-MongoDB Course API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Get a paginated list of all courses
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of courses
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course (requires authentication)
      parameters:
      - description: Course details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.CreateCourseDto'
      produces:
      - application/json
      responses:
        "201":
          description: Course created successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.Course'
        "400":
          description: Bad request - validation error or duplicate course
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new course
      tags:
      - courses
...
  /users:
    get:
      description: Get a paginated list of all users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of users
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with name, email, password and optional role
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.UserResponse'
        "400":
          description: Bad request - validation error or user already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.User'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.User'
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - users
  /users/drop:
    delete:
      description: Delete all users from the database (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: All users deleted successfully
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Drop user collection
      tags:
      - users
  /users/me:
    get:
      description: Get the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"